third_party/WebKit/LayoutTests/http/tests/credentialmanager/publickeycredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: PublicKeyCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

promise_test(async _ => {
   assert_equals(await PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable(), false);
}, "isUserVerifyingPlatformAuthenticatorAvailable() is false by default.");

</script>
